Move functions and hintings to new files and add bucket arguments.
